+++ pygccxml/declarations/calldef.py
@@ -738,7 +738,7 @@ class constructor_t(member_calldef_t):
     def is_copy_constructor(self):
         """
         Returns True if described declaration is copy constructor,
-        otherwise False
+        otherwise False.
 
         """
 
@@ -753,6 +753,17 @@ class constructor_t(member_calldef_t):
         # We have only one argument, get it
         arg = args[0]
 
+        if not isinstance(arg.type, cpptypes.compound_t):
+            # An argument of type declarated_t (a typedef) could be passed to
+            # the constructor; and it could be a reference.
+            # But in c++ you can NOT write :
+            #    "typedef class MyClass { MyClass(const MyClass & arg) {} }"
+            # If the argument is a typedef, this is not a copy constructor.
+            # See the hierarchy of declarated_t and coumpound_t. They both
+            # inherit from type_t but are not related so we can discriminate
+            # between them.
+            return False
+
         # The argument needs to be passed by reference in a copy constructor
         if not type_traits.is_reference(arg.type):
             return False
@@ -764,8 +775,14 @@ class constructor_t(member_calldef_t):
         un_aliased = type_traits.remove_alias(arg.type.base)
         # un_aliased now refers to const_t instance
         if not isinstance(un_aliased.base, cpptypes.declarated_t):
+            # We are looking for a declaration
+            # If "class MyClass { MyClass(const int & arg) {} }" is used,
+            # this is not copy constructor, so we return False here.
+            # -> un_aliased.base == cpptypes.int_t (!= cpptypes.declarated_t)
             return False
 
+        # Final check: compare the parent (the class declaration for example)
+        # with the declaration of the type passed as argument.
         return id(un_aliased.base.declaration) == id(self.parent)
 
     @property