+++ unittests/data/test_copy_constructor.hpp
@@ -1,0 +1,21 @@
+// Copyright 2014-2016 Insight Software Consortium.
+// Distributed under the Boost Software License, Version 1.0.
+// See http://www.boost.org/LICENSE_1_0.txt
+
+typedef const int & myvar;
+
+class test
+{
+    public:
+        // The copy constructor
+        test(const test & t0){};
+
+        // A constructor
+        test(const float & t0){};
+
+        // A constructor with a typedef
+        test(myvar t0){};
+};
+
+// An empty class; C++ will automatically create a copy constructor
+class test2 {};