+++ docs/history.rst
@@ -11,6 +11,13 @@ to python 3 (keeping it compatible with python 2).
 In Mai 2014, Michka Popoff and the Insight Software Consortium revived pygccxml
 by setting up a git repositery on github, hosted along with gccxml.
 
+Version 1.7.2 (unreleased)
+--------------------------
+
+1. Fix exception in is_copy_constructor() when the constructor's argument was
+   a typedef. is_copy_constructor() will now return False instead of failing.
+   See issue #27.
+
 Version 1.7.1
 -------------